move functions out of the consts.go file
